add eval script

+++ eval_alexnet.py
@@ -122,7 +122,7 @@ def main(args):
         with torch.cuda.amp.autocast():
             with torch.no_grad():
                 batch_logits_test_all = classifier_fc(classifier_embed(test_images))
-    
+        batch_logits_test_all = batch_logits_test_all.float()
         batch_y_pred_all = batch_logits_test_all.argmax(dim=1)
         y_test_pred_all.append(batch_y_pred_all.cpu())
         print(batch_logits_test_all.sum(1))
